Changes<,> - add parameter change to .ApplyChangesTo()

+++ Json/Fliox.Hub/Client/Event/Changes.cs
@@ -138,25 +138,29 @@ namespace Friflo.Json.Fliox.Hub.Client
         }
         
         /// <summary> Apply the container changes to the given <paramref name="entitySet"/> </summary>
-        public void ApplyChangesTo(EntitySet<TKey, T> entitySet) {
+        public void ApplyChangesTo(EntitySet<TKey, T> entitySet, Change change = ChangeFlags.All) {
             if (Count == 0)
                 return;
             var client = entitySet.intern.store;
             using (var pooled = client._intern.pool.ObjectMapper.Get()) {
                 var mapper          = pooled.instance;
                 var localCreates    = rawCreates;
-                if (localCreates.Count > 0) {
+                if ((change & Change.create) != 0 && localCreates.Count > 0) {
                     var entityKeys = GetKeysFromEntities (client, entitySet.GetKeyName(), localCreates);
                     SyncPeerEntities(entitySet, entityKeys, localCreates, mapper);
                 }
                 var localUpserts    = rawUpserts;
-                if (localUpserts.Count > 0) {
+                if ((change & Change.upsert) != 0 && localUpserts.Count > 0) {
                     var entityKeys = GetKeysFromEntities (client, entitySet.GetKeyName(), localUpserts);
                     SyncPeerEntities(entitySet, entityKeys, localUpserts, mapper);
                 }
-                entitySet.PatchPeerEntities(Patches, mapper);
+                if ((change & Change.patch) != 0) {
+                    entitySet.PatchPeerEntities(Patches, mapper);
+                }
             }
-            entitySet.DeletePeerEntities(Deletes);
+            if ((change & Change.delete) != 0) {
+                entitySet.DeletePeerEntities(Deletes);
+            }
         }
         
         private static List<JsonKey> GetKeysFromEntities(FlioxClient client, string keyName, List<JsonValue> entities) {
